requirements.txt
# dev
black==21.7b0
flake8==3.9.2
isort==5.9.3
mypy==0.910
pytest==6.2.4
pytest-cov==2.12.1
pytest-mock==3.6.1

# Google Colaboratory
google-cloud-bigquery==1.21.0
ipython==5.5.0
pandas==1.1.5
